Add preliminary JS FFI support
This required making Env something that can be partially shared for JS
callbacks to work. As a binding-storing member had to be put behing
Rc<RefCell<_>> for that to work, Env::get_binding can no longer return
Cow<'a, Val>, as 'a would have to be the lifetime of a local borrow.
This will be addressed later on, when the design somewhat stabilizes.